Renaming server and services
