+++ neurostack/start_server.py
@@ -1,8 +1,8 @@
-from server.p300_server import P300Service
+from server.server import NeurostackServer
 
 
 # Run neurostack server on localhost:8001
-service = P300Service()
+service = NeurostackServer()
 service.initialize_handlers()
 
 service.app.run(host='localhost', port=8001)
